Conflicts:
	ManagerM/Makefile
	ManagerM/Makefile.Debug
	ManagerM/Makefile.Release
	ManagerM/ManagerM.pro
	ManagerM/ManagerM.pro.user
	ManagerM/debug/ManagerM.exe
	ManagerM/debug/moc_generado.cpp
	ManagerM/debug/moc_lineedit.cpp
	ManagerM/debug/moc_search_articulo.cpp
	ManagerM/debug/moc_ui_advanced_search_articulo.cpp
	ManagerM/debug/moc_ui_colaborador.cpp
	ManagerM/debug/moc_ui_edit_articulo.cpp
	ManagerM/debug/moc_ui_edit_boleta_venta.cpp
	ManagerM/debug/moc_ui_edit_compra_orden.cpp
	ManagerM/debug/moc_ui_edit_personalinformation.cpp
	ManagerM/debug/moc_ui_edit_proveedor.cpp
	ManagerM/debug/moc_ui_login.cpp
	ManagerM/debug/moc_ui_main.cpp
	ManagerM/debug/moc_ui_module_articulos.cpp
	ManagerM/debug/moc_ui_module_clientes.cpp
	ManagerM/debug/moc_ui_module_colaboradores.cpp
	ManagerM/debug/moc_ui_module_compras.cpp
	ManagerM/debug/moc_ui_module_home.cpp
	ManagerM/debug/moc_ui_module_ventas.cpp
	ManagerM/debug/moc_ui_new_articulo.cpp
	ManagerM/debug/moc_ui_new_boleta_sistema.cpp
	ManagerM/debug/moc_ui_new_compra_orden.cpp
	ManagerM/debug/moc_ui_new_factura_compra.cpp
	ManagerM/debug/moc_ui_new_factura_sistema.cpp
	ManagerM/debug/moc_ui_new_flete.cpp
	ManagerM/debug/moc_ui_new_guiarr.cpp
	ManagerM/debug/moc_ui_new_marca.cpp
	ManagerM/debug/moc_ui_new_medida.cpp
	ManagerM/debug/moc_ui_new_orden.cpp
	ManagerM/debug/moc_ui_new_proveedor.cpp
	ManagerM/debug/moc_ui_new_transportista.cpp
	ManagerM/debug/moc_ui_new_venta.cpp
	ManagerM/debug/moc_ui_new_venta_boleta.cpp
	ManagerM/debug/moc_ui_new_venta_cotizacion.cpp
	ManagerM/debug/moc_ui_new_venta_factura.cpp
	ManagerM/debug/moc_ui_new_venta_proforma.cpp
	ManagerM/debug/moc_ui_newedit_cliente.cpp
	ManagerM/debug/moc_ui_newedit_colaborador.cpp
	ManagerM/debug/moc_ui_opciones.cpp
	ManagerM/debug/moc_ui_opciones_articulo.cpp
	ManagerM/debug/moc_ui_opciones_banco.cpp
	ManagerM/debug/moc_ui_opciones_base_datos.cpp
	ManagerM/debug/moc_ui_opciones_comprobante.cpp
	ManagerM/debug/moc_ui_opciones_general.cpp
	ManagerM/debug/moc_ui_opciones_proveedor.cpp
	ManagerM/debug/moc_ui_opciones_tienda.cpp
	ManagerM/debug/moc_ui_tool_bar.cpp
	ManagerM/debug/qrc_Icons.cpp
	ManagerM/managerm.log
	ManagerM/object_script.ManagerM.Debug
	ManagerM/object_script.ManagerM.Release
	ManagerM/share_include.h
	ManagerM/sistema.cpp
	ManagerM/ui_generado.h
	ManagerM/ui_new_venta.cpp
	ManagerM/ui_search_articulo.h
	ManagerM/ui_ui_advanced_search_articulo.h
	ManagerM/ui_ui_colaborador.h
	ManagerM/ui_ui_edit_articulo.h
	ManagerM/ui_ui_edit_boleta_venta.h
	ManagerM/ui_ui_edit_compra_orden.h
	ManagerM/ui_ui_edit_personalinformation.h
	ManagerM/ui_ui_edit_proveedor.h
	ManagerM/ui_ui_login.h
	ManagerM/ui_ui_main.h
	ManagerM/ui_ui_module_articulos.h
	ManagerM/ui_ui_module_clientes.h
	ManagerM/ui_ui_module_colaboradores.h
	ManagerM/ui_ui_module_compras.h
	ManagerM/ui_ui_module_home.h
	ManagerM/ui_ui_module_ventas.h
	ManagerM/ui_ui_new_articulo.h
	ManagerM/ui_ui_new_boleta_sistema.h
	ManagerM/ui_ui_new_compra_orden.h
	ManagerM/ui_ui_new_factura_compra.h
	ManagerM/ui_ui_new_factura_sistema.h
	ManagerM/ui_ui_new_flete.h
	ManagerM/ui_ui_new_guiarr.h
	ManagerM/ui_ui_new_marca.h
	ManagerM/ui_ui_new_medida.h
	ManagerM/ui_ui_new_orden.h
	ManagerM/ui_ui_new_proveedor.h
	ManagerM/ui_ui_new_transportista.h
	ManagerM/ui_ui_new_venta.h
	ManagerM/ui_ui_new_venta_boleta.h
	ManagerM/ui_ui_new_venta_factura.h
	ManagerM/ui_ui_newedit_cliente.h
	ManagerM/ui_ui_newedit_colaborador.h
	ManagerM/ui_ui_opciones.h
	ManagerM/ui_ui_opciones_articulo.h
	ManagerM/ui_ui_opciones_banco.h
	ManagerM/ui_ui_opciones_base_datos.h
	ManagerM/ui_ui_opciones_comprobante.h
	ManagerM/ui_ui_opciones_general.h
	ManagerM/ui_ui_opciones_proveedor.h
	ManagerM/ui_ui_opciones_tienda.h
	ManagerM/ui_ui_tool_bar.h

+++ ManagerM/ui_newedit_cliente.cpp
@@ -31,7 +31,7 @@ void ui_newedit_cliente::llenarComboBox()
     ui->comboBox_region->clear();
     vector<_QSTR> regiones=SYSTEM->getAllRegiones();
 
-    for(int i=0;i<regiones.size();i++)
+    for(int i=0;i<(int)regiones.size();i++)
         ui->comboBox_region->addItem(QString(regiones[i]));
 }
